requirements.txt
# Version-pinned dependencies for production stability
# Pin major versions to avoid breaking changes

flask==3.0.0
pdfplumber==0.11.0
python-dotenv==1.0.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.11
rapidfuzz==3.6.1
scikit-learn==1.4.0
reportlab==4.1.0
werkzeug==3.0.1
sentence-transformers==2.3.1
spacy==3.7.4
redis==5.0.1
pyjwt==2.8.0
bcrypt==4.1.2
python-docx==1.1.0
requests==2.31.0
alembic==1.13.2
